requirements.txt
streamlit
sqlalchemy>=2
psycopg2-binary
pandas
python-dotenv
plotly
bcrypt
